Updated getByPriority

+++ src/main/java/com/example/ToDoList/DAO/PriorityRepository.java
@@ -3,5 +3,5 @@ package com.example.ToDoList.DAO;
 import com.example.ToDoList.entity.Priority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PriorityRepository extends JpaRepository<Priority, Integer> {
+public interface PriorityRepository extends JpaRepository<Priority, Long> {
 }
